# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "이메일·비밀번호로 로그인, access/refresh 토큰 발급",
        "operationId": "JammitAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JammitLoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공 시 CommonResponseDto, result에 토큰·유저 정보",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "code": 200,
                    "message": "OK",
                    "result": {
                      "accessToken": "…",
                      "refreshToken": "…",
                      "user": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "이메일 또는 비밀번호 불일치"
          }
        },
        "summary": "로그인",
        "tags": [
          "Jammit Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "refresh token으로 새 access token 발급",
        "operationId": "JammitAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JammitRefreshRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CommonResponseDto 래핑 결과"
          },
          "401": {
            "description": "유효하지 않은 refresh token"
          }
        },
        "summary": "토큰 갱신",
        "tags": [
          "Jammit Auth"
        ]
      }
    },
    "/auth/email/send-code": {
      "post": {
        "description": "6자리 코드를 메일로 전송 (개발/스테이징에서 로그 확인 가능할 수 있음)",
        "operationId": "JammitAuthController_sendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JammitEmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "발송 완료 (CommonResponseDto)"
          }
        },
        "summary": "이메일 인증 코드 발송",
        "tags": [
          "Jammit Auth"
        ]
      }
    },
    "/auth/email/verify-code": {
      "post": {
        "description": "코드 일치 시 200, 실패 시 401 등",
        "operationId": "JammitAuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JammitVerifyCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "검증 성공 (HTTP status는 응답 본문과 함께 설정됨)"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "코드 불일치 또는 만료"
          }
        },
        "summary": "이메일 인증 코드 검증",
        "tags": [
          "Jammit Auth"
        ]
      }
    },
...
    "/comment": {
      "get": {
        "description": "영상별 페이지네이션",
        "operationId": "CommentController_getComment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "query",
            "description": "영상 UUID",
            "schema": {
              "example": "xxx-uuid",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "페이지 (1부터)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "페이지네이션 객체 (data: 댓글 항목 배열)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "totalCount": 0,
                    "totalPage": 1,
                    "page": 1,
                    "data": [],
                    "message": "영상에 달린 댓글 리스트를 조회했습니다."
                  }
                }
              }
            }
          },
          "400": {
            "description": "조회 실패(잘못된 요청 등)"
          },
          "500": {
            "description": "서버 오류"
          }
        },
        "summary": "댓글 목록",
        "tags": [
          "Comment"
        ]
      },
      "post": {
        "operationId": "CommentController_createComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "작성된 댓글 요약"
          },
          "401": {
            "description": "미인증"
          },
          "404": {
            "description": "영상 없음"
          },
          "500": {
            "description": "댓글 쓰기 중 오류"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "댓글 작성",
        "tags": [
          "Comment"
        ]
      }
    }
  },
  "info": {
    "title": "jammit api문서",
    "description": "jammit 개발을 위한 API 문서입니다.\n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT access token",
        "in": "header"
      }
    },
    "schemas": {
      "JammitLoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "로그인 이메일"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "비밀번호"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "JammitRefreshRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "발급받은 JWT refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "JammitEmailRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "인증 코드를 받을 이메일"
          }
        },
        "required": [
          "email"
        ]
      },
      "JammitVerifyCodeRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "이메일로 받은 6자리 코드"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "example": "b1a8f3e1-3b0f-4e9b-98a2-c4f0e6d3a3b4",
            "description": "영상 UUID"
          },
          "content": {
            "type": "string",
            "example": "탕탕후루후루 탕탕 후루루루루",
            "description": "댓글내용"
          }
        },
        "required": [
          "videoId",
          "content"
        ]
      }
    }
  }
}